scan_final/requirements.txt
django-jsonfield==0.9.15
